pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "numpy", "numba"]
build-backend = "setuptools.build_meta"

[project]
name = "raw_alchemy_studio"
version = "0.3.5"
description = "Raw Alchemy Studio - A fast culling/grading tool that applies cinematic LUTs to RAW photos with mathematical precision."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "requests",
    "tifffile",
    "click",
    "rawpy @ git+https://github.com/shenmintao/rawpy.git",
    "colour-science @ git+https://github.com/colour-science/colour.git@develop",
    "scipy",
    "Pillow",
    "pillow-heif",
    "pyinstaller",
    "zstandard",
    "ordered-set",
    "numba",
    "matplotlib",
    "PySide6-Fluent-Widgets",
    "Send2Trash",
    "packaging",
    "loguru",
    "pyexiv2"
]

[project.urls]
"Homepage" = "https://github.com/shenmintao/raw-alchemy"
"Bug Tracker" = "https://github.com/shenmintao/raw-alchemy/issues"

[project.scripts]
raw-alchemy = "raw_alchemy.cli:main"
raw-alchemy-gui = "raw_alchemy.main:main"
